Refactor Filer to Storage
